+++ pipe/bypass.go
@@ -1,6 +1,6 @@
 package pipe
 
-import "github.com/mariomac/pipe/internal/connect"
+import "github.com/mariomac/pipes/pipe/internal/connect"
 
 // IgnoreStart is a convenience function to explicitly specify that the returned StartFunc
 // is going to be ignored/bypassed by the pipes library
