+++ pipe/internal/connect/connectors_test.go
