+++ pipe/options.go
