+++ pipe/runner.go
@@ -1,0 +1,6 @@
+package pipe
+
+type Runner struct {
+	startNodes    []startable
+	terminalNodes []doneable
+}
